requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
beartype==0.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
CoLT5-attention==0.11.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
einops==0.8.1
einx==0.3.0
filelock==3.17.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.3
hyper-connections==0.1.15
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
local-attention==1.11.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.0
pyarrow==19.0.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3

